feat: introduce the TwitterHeader component
