+++ src/app/home/[username]/page.tsx
@@ -1,10 +1,14 @@
+import TwitterHeader from "@/components/Header/TwitterHeader"
 
 function Page(
     { params: { username } }
         :
         { params: { username: string } }) {
     return (
-        <div>{username} Page</div>
+        <section className='w-1/3 grow'>
+            <TwitterHeader section="Profile" />
+            <div>{username} Page</div>
+        </section>
     )
 }
